Refactor the daylights out of queries.py to make it editable
