+++ indicators/queries/__init__.py
@@ -1,0 +1,6 @@
+from program_queries import ProgramWithMetrics
+from iptt_queries import IPTTIndicator
+from indicators_queries import (
+    ResultsIndicator,
+    MetricsIndicator
+)
